+++ package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "cli": "node ./src/cli/cli.js  --value1 1 --value2 2"
+    "cli": "node ./src/cli/cli.js  --value1 1 --value2 2",
+    "modules": "node --experimental-json-modules ./src/modules/cjsToEsm.mjs"
   },
   "repository": {
     "url": "git+https://github.com/AlreadyBored/node-nodejs-basics.git"
